+++ IcuLib/wintz.cpp
@@ -368,7 +368,7 @@ uprv_detectWindowsTimeZone()
      */
     if (tmpid[0] != 0) 
     {
-        len = uprv_strlen(tmpid);
+        len = (int32_t)uprv_strlen(tmpid);
         icuid = (char*)uprv_calloc(len + 1, sizeof(char));
         if (icuid != NULL) 
         {
